Dev : refactored big
