+++ packages/arb-token-bridge-ui/jest.config.js
@@ -5,7 +5,7 @@ const createJestConfig = nextJest({
 })
 
 const customJestConfig = {
-  testEnvironment: 'jest-environment-jsdom'
+  testEnvironment: 'node'
 }
 
 // createJestConfig is exported this way to ensure that next/jest can load the Next.js config which is async
