+++ packages/arb-token-bridge-ui/src/util/deposits/helpers.ts
@@ -9,9 +9,7 @@ import {
   L1ToL2MessageReaderClassic
 } from '@arbitrum/sdk/dist/lib/message/L1ToL2Message'
 import { Provider } from '@ethersproject/providers'
-import { AssetType } from '../hooks/arbTokenBridge.types'
-import { Transaction } from '../hooks/useTransactions'
-import { getL1TokenData } from '../util'
+import { AssetType, getL1TokenData, Transaction } from 'token-bridge-sdk'
 
 export const updateAdditionalDepositData = async (
   depositTx: Transaction,
