+++ packages/arb-token-bridge-ui/src/util/deposits/tests/fetchDepositsFromSubgraph.test.ts
@@ -1,7 +1,3 @@
-/**
- * @jest-environment jsdom
- */
-
 import {
   getQueryCoveringClassicOnlyWithoutResults,
   getQueryCoveringClassicOnlyWithResults,
