+++ packages/token-bridge-sdk/src/util/index.ts
@@ -162,13 +162,3 @@ export function isClassicL2ToL1TransactionEvent(
 ): event is EventArgs<ClassicL2ToL1TransactionEvent> {
   return typeof (event as any).batchNumber !== 'undefined'
 }
-
-export const sanitizeQueryParams = (data: any) => {
-  return JSON.parse(JSON.stringify(data))
-}
-
-export const getAPIBaseUrl = () => {
-  // if dev environment, eg. tests, then prepend actual running environment
-  // Resolves: next-js-error-only-absolute-urls-are-supported in test:ci:sdk
-  return process.env.NODE_ENV === 'test' ? 'http://localhost:3000' : ''
-}
