--- packages/token-bridge-sdk/src/withdrawals/index.ts
@@ -1,3 +1,0 @@
-export { fetchETHWithdrawalsFromEventLogs } from './fetchETHWithdrawalsFromEventLogs'
-export { fetchTokenWithdrawalsFromEventLogs } from './fetchTokenWithdrawalsFromEventLogs'
-export { fetchWithdrawals } from './fetchWithdrawals'
